% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Tax.R
\name{summary.Tax}
\alias{summary.Tax}
\title{Summarize taxonomic information}
\usage{
\method{summary}{Tax}(data, columnsIndex)
}
\arguments{
\item{data}{A \code{"data.frame"} object.}

\item{columnIndex}{A numeric vector of length 1 to 4 indicating the position of the Species, Genus, and Family columns in the data frame. By default \code{"columnIndex = c(5, 4, 3)"} according to the table structure of the \code{websearch.Tax()} output.}
}
\description{
\code{summary.Tax} summarize the taxonomic information contained in a table.
}
