% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNames.Tax.R
\name{updateNames.Tax}
\alias{updateNames.Tax}
\title{Update taxonomic names}
\usage{
updateNames.Tax(names, validationMatrix, threshold = 0.8, markNoUpdated = TRUE)
}
\arguments{
\item{names}{A vector of type \code{"character"} with the input names to be updated.}

\item{validationMatrix}{A \code{"data.frame"} object obtained from \link[taxTools]{validate.Tax} function.}

\item{threshold}{A number between (0, 1] indicating the proportion of similitude to be considered as the cutoff value for update the input names. In case the validation matrix contains multi-match and more than one name agrees with the threshold criteria, the update will be made using the name with the highest score.}

\item{markNoUpdated}{Boolean indicating if the names that won't be updated must be marked to identify them. When \code{markNoUpdated = TRUE} an * will be added at the end of the name, on the contrary, the names with scores lower than threshold value will keep as is.}
}
\description{
\code{updateNames.Tax} updates the user's input taxonomic names based on a validation matrix.
}
