% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websearch.Tax.R
\name{websearch.Tax}
\alias{websearch.Tax}
\title{Download the most recent taxonomy from online databases}
\usage{
websearch.Tax(class, source = c("AMNH", "AB", "TRDB"), country)
}
\arguments{
\item{class}{A string of type \code{"character"} indicating the taxonomic Class of interest. Available ones are: \code{"Amphibia"} or shorter \code{"Amp"} for amphibians, and \code{"Reptilia"} or shorter \code{"Rep"} for reptiles.}

\item{source}{A string of type \code{"character"} indicating the source for retrieving the taxonomy. Available ones are: \code{"AMNH"} (\href{https://amphibiansoftheworld.amnh.org}{Amphibians of the World}) and \code{"AB"} (\href{https://amphibiaweb.org}{AmphibiaWeb}) for amphibians, and \code{"TRDB"} (\href{http://www.reptile-database.org}{The Reptile Database}) for reptiles.}

\item{country}{A string or a vector of type \code{"character"} indicating the country or countries for filtering the search. Country names must be written as in the reference sources search options (see each source website).}
}
\value{
\code{websearch.Tax} returns a \code{"data.frame"} object with the most recent taxonomy for the selected Class, including Order, Class, Family, Genus, Species, authorship (author + year), and country.
}
\description{
\code{websearch.Tax} retrieves the most recent taxonomic classification for a given Class from online reference databases. This function requires an active internet connection.

\strong{Note:} Currently, this function is only supported for amphibians and reptiles.
}
\details{
This function uses the web scraping method for retireve the data, therefore, a large number of queries, or queries for many countries will take more time or may temporary block the access to the website.

Given the script for each source depends on the HTML structure of each webpage, any change in its structure may be the queries to stop working suddenly. In this last case, please email the package author.
}
